Split PyInterpreter into its own file.
I also took the opportunity to update the documentation a little
for clarity.


[ghstack-poisoned]
